+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Changed
+- Switching focus is now prevented while a context menu is active.
+
 ## [0.2.0]
 ### Added
 - Make focus change delay configurable.